pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "woolf-intertextuality"
version = "0.1.0"
description = "Detecting intertextuality between Mrs Dalloway and The Odyssey using RAG"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "qdrant-client",
    "openai",
    "numpy",
    "pandas",
    "matplotlib",
    "nltk",
    "pydantic-settings>=2.6.1",
    "pydantic>=2.9.2",
    "qdrant-haystack>=7.0.0",
    "haystack-ai>=2.7.0",
    "rich>=13.9.4",
    "jinja2>=3.0.0",
    "scikit-learn>=1.6.0",
    "krippendorff>=0.8.0",
    "tiktoken>=0.8.0",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.3.0",
]

[tool.pytest.ini_options]
pythonpath = [
    "."
]
testpaths = [
    "tests"
]

# Add these sections for build configuration
[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.9
target-version = "py39"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Use double quotes for strings.
quote-style = "double"

# Indent with spaces, rather than tabs.
indent-style = "space"

# Respect magic trailing commas.
skip-magic-trailing-comma = false

# Automatically detect the appropriate line ending.
line-ending = "auto"
